pointerevents/pointerevent_pointerEnter.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: Dispatch pointerEnter. </title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="When a pointing device is moved into the hit test boundaries of an element or one of its descendants, the pointerenter event must be dispatched."/>

        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
        <script type="text/javascript">

            setup({ explicit_done: true });

            var detected_pointertypes = {};
            add_completion_callback(showPointerTypes);

            var target0;
            var test_pointerEnter;

            function run() {
                target0 = document.getElementById("target0");
                test_pointerEnter = async_test("pointerenter event");  // set up test harness

                on_event(target0, "pointerenter", function (event) {
                    check_PointerEvent(event);
                    test_pointerEnter.step(function () {
                        assert_equals(event.type, "pointerenter", "pointer event received" + event.type);
                    });
                    detected_pointertypes[event.pointerType] = true;
                    test_pointerEnter.done(); // complete test
                });
            }

            function showPointerTypes() {
                var complete_notice = document.getElementById("complete-notice");
                var pointertype_log = document.getElementById("pointertype-log");
                var pointertypes = Object.keys(detected_pointertypes);
                pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
                if (test_pointerEnter.status == test_pointerEnter.NOTRUN
                    || test_pointerEnter.status == test_pointerEnter.TIMEOUT) {
                    complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                    complete_notice.style.background = "red";
                }
                complete_notice.style.display = "block";
            }
        </script>

            <style>
            div {
                margin: 0em;
                padding: 2em;
            }

            #target0 {
                background: yellow;
                border: 1px solid orange;
            }

            #complete-notice {
                border: 1px solid #0a0;
                display: none;
            }

            #pointertype-log {
                font-weight: bold;
            }
        </style>
    </head>
    <body onload="run()">
        <h1>Pointer Event: Dispatch pointerEnter</h1>
        <h4> 
            Test Description: 
            When a pointing device is moved into the hit test boundaries of an element or one of its descendants, the pointerenter event must be dispatched. 
        </h4>
        <div id="target0">
            Use the mouse or pen to move over this box.
        </div>
        <div id="complete-notice">
            <p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
            <p>Refresh the page to run the tests again with a different pointer type?</p>
        </div>

        <div id="log"></div>
    </body>
</html>
